add an update endpoint for post : update by id

+++ backend/src/posts/posts.service.ts
@@ -1,12 +1,17 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
-import { CreatedPost, Image, PostToCreate } from './dto/post.dto';
+import {
+  ForbiddenException,
+  Injectable,
+  InternalServerErrorException,
+  NotFoundException,
+} from '@nestjs/common';
+import { PostType, Image, PostToCreate, NewPostDataDto } from './dto/post.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
 
 @Injectable()
 export class PostsService {
   constructor(private prisma: PrismaService) {}
 
-  async createPost(postToCreate: PostToCreate): Promise<CreatedPost> {
+  async createPost(postToCreate: PostToCreate): Promise<PostType> {
     const userId = postToCreate.userId;
     const user = await this.prisma.user.findUnique({
       where: { id: userId },
@@ -35,12 +40,12 @@ export class PostsService {
         where: { id: image.id },
         data: { postId: createdPost.id },
       });
-      console.log("updated image" + updatedImg);
+      console.log('updated image' + updatedImg);
     }
     return createdPost;
   }
 
-  async deletePost(userId: string, postId: string): Promise<CreatedPost> {
+  async deletePost(userId: string, postId: string): Promise<PostType> {
     const user = await this.prisma.user.findUnique({
       where: { id: userId },
     });
@@ -52,6 +57,9 @@ export class PostsService {
     });
     if (!post) {
       throw new NotFoundException('User not found');
+    }
+    if(post.userId !== userId) {
+      throw new ForbiddenException('Only the original author can delete his content!')
     }
     try {
       const deletedPost = await this.prisma.post.delete({
@@ -70,6 +78,71 @@ export class PostsService {
       };
     } catch (error) {
       console.log('DELETE POST ERROR: ' + error);
+    }
+  }
+
+  async updatePost(
+    userId: string,
+    postId: string,
+    newPostData: NewPostDataDto,
+  ): Promise<PostType> {
+    const user = await this.prisma.user.findUnique({
+      where: { id: userId },
+    });
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+    const post = await this.prisma.post.findUnique({
+      where: { id: postId },
+    });
+    if (!post) {
+      throw new NotFoundException('Post not found');
+    }
+    if(post.userId !== userId) {
+      throw new ForbiddenException('Only the original author can edit his content!')
+    }
+    console.log(newPostData);
+    
+    if (Object.keys(newPostData).length === 0 ) {
+      throw new NotFoundException('Nothing to update');
+    }
+    try {
+
+      if(newPostData.picturePath){
+        const updatedImg = await this.prisma.image.update({
+          where:{
+            path:post.picturePath
+          },
+          data:{path:newPostData.picturePath}
+        })
+        console.log("Updated Image" + updatedImg.path);
+        
+      }
+
+      const updatedPostData = {
+        ...post,
+        ...newPostData,
+      };
+
+      const updatedPost = await this.prisma.post.update({
+        where: { id: postId },
+        data: updatedPostData,
+      });
+ 
+      const returnedUpdatedPost = {
+        id: updatedPost.id,
+        userId: updatedPost.userId,
+        title: updatedPost.title,
+        description: updatedPost.description,
+        location: updatedPost.location,
+        picturePath: updatedPost.picturePath,
+        userpicturePath: updatedPost.userpicturePath,
+      };
+      return returnedUpdatedPost;
+    } catch (error) {
+      console.log('UPDATE POST ERROR: ' + error);
+
+      throw new InternalServerErrorException('Failed to update user');
     }
   }
 }
